hack_mercari/backend/requirements.txt
celery
django
djangorestframework
django_cors_headers
pusher